Committing Final Project Files
This commit includes the final project iPython Notebook that contains the code to install the package, as well as an example of simple image processing and figure production. It also includes 2 .tiff files that are the dataset used in the code. Finally it includes the pdf report about the Scikit-Image package